wip: pass task's authorization to query system, if using sessions
The query system specifically expects an Authorization. When a request
comes in using a Session, use the target task's Authorization, if we are
allowed to read it, when executing a query against the system bucket.

Reason for WIP: the task's created authorization doesn't have permission
to read itself, but it should be able to. Also, the task validator
expects an authorization to read any task in the org, but it should only
check the task with the specific ID, when possible.